+++ contracts/core/AlcorPoolDeployer.sol
@@ -18,7 +18,7 @@ contract AlcorPoolDeployer is IAlcorPoolDeployer {
         address token1;
         uint256 expiration;
         uint160 optionStrikePriceX96;
-        uint8 poolFee;
+        uint32 poolFee;
         int24 tickSpacing;
     }
 
@@ -37,7 +37,7 @@ contract AlcorPoolDeployer is IAlcorPoolDeployer {
         address token0,
         address token1,
         uint256 optionExpiration,
-        uint8 poolFee,
+        uint32 poolFee,
         uint160 optionStrikePriceX96,
         int24 tickSpacing
     ) internal virtual returns (address pool) {
